requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
Django==3.1
django-crispy-forms==1.9.2
django-debug-toolbar==2.2
django-modelcluster==5.0.2
django-object-actions==3.0.1
django-qr-code==1.2.0
django-sendsms==0.4
django-taggit==1.3.0
django-treebeard==4.3.1
djangocms-installer==1.2.3
djangorestframework==3.11.1
draftjs-exporter==4.1.0
html5lib==1.1
idna==2.10
l18n==2018.5
Pillow==7.2.0
pkg-resources==0.0.0
pytz==2020.1
qrcode==6.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
tzlocal==2.1
Unidecode==1.1.1
urllib3==1.25.10
webencodings==0.5.1
Willow==1.4
XlsxWriter==1.3.3
